build/windows: Simplify installer .PS1 script
This adds a bit more verbosity to the installer script while omitting some.

The version params (e.g. gimp_version) are dropped since now config.h is
mandatory (this is a natural conclusion of the generation of assets).

+++ build/windows/installer/3_dist-gimp-inno_sym.sh
@@ -17,17 +17,19 @@ for ARTIFACTS_SUFFIX in "${archsArray[@]}"; do
     NAME="${bin##*/}"
     DIR="${debug%/*}/.debug/"
 
-    ## Split/extract DWARF symbols from binary to .debug
-    echo "(INFO): extracting DWARF symbols from $NAME to $DIR"
-    objcopy --only-keep-debug $bin $debug
+    if [ ! -f "$DIR/$NAME.debug" ]; then
+      ## Split/extract DWARF symbols from binary to .debug
+      echo "(INFO): extracting DWARF symbols from $NAME to $DIR"
+      objcopy --only-keep-debug $bin $debug
 
-    ## Link .debug to binary
-    objcopy --add-gnu-debuglink=$debug $bin --strip-unneeded
+      ## Link .debug to binary
+      objcopy --add-gnu-debuglink=$debug $bin --strip-unneeded
 
-    ## Move .debug files to .debug folder
-    if [ ! -d "$DIR" ]; then
-      mkdir "$DIR"
+      ## Move .debug files to .debug folder
+      if [ ! -d "$DIR" ]; then
+        mkdir "$DIR"
+      fi
+      mv "$debug" "$DIR"
     fi
-    mv "$debug" "$DIR"
   done
 done
